move some functions from AdminLogQuery to GaeVersionApi
