+++ src/main/java/teammates/common/util/GaeVersionApi.java
@@ -11,6 +11,15 @@ import com.google.appengine.api.modules.ModulesServiceFactory;
  * A wrapper class for GAE application admin API. e.g. version management
  */
 public class GaeVersionApi {
+    /**
+     * Maximum number of versions to query.
+     * The current value will include the current version and its 5 preceding versions.
+     */
+    private static final int MAX_VERSIONS_TO_QUERY = 6;
+    
+    /**
+     * Default constructor.
+     */
     public GaeVersionApi() {
     }
     
@@ -36,4 +45,36 @@ public class GaeVersionApi {
         return new Version(modulesService.getCurrentVersion());
     }
     
+    /**
+     * Gets a few recent versions for log query.
+     * @return a list of default versions for query.
+     */
+    public List<String> getDefaultVersionIdsForLogQuery() {
+        List<Version> versionList = getAvailableVersions();
+        Version currentVersion = getCurrentVersion();
+        
+        List<String> defaultVersions = new ArrayList<String>();
+        try {
+            int currentVersionIndex = versionList.indexOf(currentVersion);
+            defaultVersions = getNextFewVersions(versionList, currentVersionIndex);
+        } catch (IndexOutOfBoundsException  e) {
+            defaultVersions.add(currentVersion.toStringForQuery());
+            Utils.getLogger().severe(e.getMessage());
+        }
+        return defaultVersions;
+    }
+
+    /**
+     * Finds at most MAX_VERSIONS_TO_QUERY nearest versions.
+     * @param currentVersionIndex starting position to get versions to query
+     */
+    private List<String> getNextFewVersions(List<Version> versionList, int currentVersionIndex) {
+        int endIndex = Math.min(currentVersionIndex + MAX_VERSIONS_TO_QUERY, versionList.size());
+        List<Version> versionSubList = versionList.subList(currentVersionIndex, endIndex);
+        List<String> versionListInString = new ArrayList<String>();
+        for(Version version : versionSubList) {
+            versionListInString.add(version.toStringForQuery());
+        }
+        return versionListInString;
+    }
 }
